updated boxes to use immutable js
