+++ src/App.js
@@ -1,20 +1,21 @@
 import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
+import {connect} from 'react-redux';
+import SplitPane from 'react-split-pane';
+import { Map, List } from 'immutable';
+
 import Grid from './components/Grid';
 import TrashCan from './components/TrashCan';
 import Window from './components/Window';
 import AddOptions from './components/AddOptions';
 import BottomOptions from './components/BottomOptions';
 import Patterns from './components/Patterns';
+import CodeModal from './components/codemirror_modal';
+import Code from './components/Codemirror';
 
-import { setBox, addBox, removeBox, setParent, addChild, removeParent, removeChild, copyBox } from './reducers/boxes';
+import { setBox, addBox, removeBox, setParent, addChild, removeParent, removeChild, copyBox } from './constants_actioncreators/boxes';
 //import { findSibling } from './reducers/siblngReducer2';
 
-
-import {connect} from 'react-redux';
-import CodeModal from './components/codemirror_modal';
-import SplitPane from 'react-split-pane';
-import Code from './components/Codemirror';
 import './style/css/App.css'
 
 const mapStateToProps = (state) => {
@@ -72,7 +73,7 @@ class App extends Component {
 
 	render(){
 		const boxes = this.props.boxes;
-		const boxIds = this.props.boxIds;
+		const boxIds = this.props.boxes.keys();
 
 		return (
 			<div className="App">
@@ -108,17 +109,17 @@ class App extends Component {
 										addChild={this.props.addChild}
 										removeParent={this.props.removeParent}
 										removeChild={this.props.removeChild}
-										id={+box}
-										x={boxes[box].x}
-										y = {boxes[box].y}
-										height={boxes[box].height}
-										width={boxes[box].width}
-										children={boxes[box].children}
-										parent={boxes[box].parent}
-										tag={boxes[box].tag}
-										css={boxes[box].css}
-										boxIds={this.props.boxIds}
-										boxes={this.props.boxes}
+										id={box}
+										x={boxes.get(box).get('x')}
+										y = {boxes.get(box).get('y')}
+										height={boxes.get(box).get('height')}
+										width={boxes.get(box).get('width')}
+										children={boxes.get(box).get('children').toJS()}
+										parent={boxes.get(box).get('parent')}
+										tag={boxes.get(box).get('tag')}
+										css={boxes.get(box).get('css')}
+										boxIds={boxIds}
+										boxes={this.props.boxes.toJS()}
 										boxCopier={this.boxCopier}
 										/>
 										)
